Rename module, add documentation to README.md
