+++ src/packages/core/equal.test.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { describe, expect, it } from "vitest";
-import { deepEqual, shallowEqual } from "./equals";
+import { deepEqual, shallowEqual } from "./equal";
 
 describe("shallowEqual", () => {
     it("returns true for identical primitives", () => {
