+++ src/packages/core/equal.ts
